+++ CHANGELOG.md
@@ -7,11 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - This `CHANGELOG.md`
+- Ruby syntax version customisable with `-r` or `--ruby-syntax`
 
 ### Fixed
 - Relative and absolute project path support
-
-### Fixed
 - typo in stale coverage warning message by @ilyakorol
 
 ## [0.1.7] - 2018-08-03